+++ scripts/systems/world/WorldEnvironmentController.gd
@@ -9,4 +9,5 @@ class_name WorldEnvironmentController
 @export var time_accelerator: TimeAccelerator
 @export var world_environment: WorldEnvironment
 @export var sun_light: DirectionalLight3D
+@export var color_grade_controller: ColorGradeController
 #@export var moon_light: DirectionalLight3D